+++ Engine/ReadNode.cpp
@@ -188,17 +188,17 @@ isGenericKnob(const std::string& knobName,
 bool
 ReadNode::isBundledReader(const std::string& pluginID)
 {
-    return boost::iequals(pluginID, PLUGINID_OFX_READOIIO) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READFFMPEG) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READPFM) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READPSD) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READKRITA) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READSVG) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READMISC) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READORA) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READCDR) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READPNG) ||
-    boost::iequals(pluginID, PLUGINID_OFX_READPDF);
+    return (pluginID == PLUGINID_OFX_READOIIO ||
+            pluginID == PLUGINID_OFX_READFFMPEG ||
+            pluginID == PLUGINID_OFX_READPFM ||
+            pluginID == PLUGINID_OFX_READPSD ||
+            pluginID == PLUGINID_OFX_READKRITA ||
+            pluginID == PLUGINID_OFX_READSVG ||
+            pluginID == PLUGINID_OFX_READMISC ||
+            pluginID == PLUGINID_OFX_READORA ||
+            pluginID == PLUGINID_OFX_READCDR ||
+            pluginID == PLUGINID_OFX_READPNG ||
+            pluginID == PLUGINID_OFX_READPDF);
 }
 
 struct ReadNodePrivate
@@ -729,7 +729,7 @@ ReadNodePrivate::refreshFileInfoVisibility(const std::string& pluginID)
     }
 
 
-    if ( hasMetaDatasKnob || ( (pluginID == PLUGINID_OFX_READFFMPEG) && hasFfprobe ) ) {
+    if ( hasMetaDatasKnob || ( ReadNode::isVideoReader(pluginID) && hasFfprobe ) ) {
         fileInfos->setSecret(false);
     } else {
         fileInfos->setSecret(true);
@@ -789,6 +789,21 @@ bool
 ReadNode::isReader() const
 {
     return true;
+}
+
+// static
+bool
+ReadNode::isVideoReader(const std::string& pluginID)
+{
+    return (pluginID == PLUGINID_OFX_READFFMPEG);
+}
+
+bool
+ReadNode::isVideoReader() const
+{
+    NodePtr p = getEmbeddedReader();
+
+    return p ? isVideoReader( p->getPluginID() ) : false;
 }
 
 bool
@@ -1117,7 +1132,7 @@ ReadNode::knobChanged(const KnobIPtr& k,
             }
         } else {
             QString ffprobePath = ReadNodePrivate::getFFProbeBinaryPath();
-            if ( (p->getPluginID() == PLUGINID_OFX_READFFMPEG) && QFile::exists(ffprobePath) ) {
+            if ( isVideoReader( p->getPluginID() ) && QFile::exists(ffprobePath) ) {
                 QProcess proc;
                 QStringList ffprobeArgs;
                 ffprobeArgs << QString::fromUtf8("-show_streams");