+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Natron multiple times on the same computer, the different processes will share t
 - fix hue display in the viewer #1554
 - EdgeBlur now has an optional Matte input used to compute the edges #1553
 - fix reading image sequences that do not have frame number 1 in them #1556
+- better/simpler GUI and documentation for the file dialog.
 
 
 ## Version 2.2.1